pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pyzbar>=0.1.9",
    "qrcode>=8.2",
]
